chore(hub): absorb aether-hub crate into aether-substrate-hub
Phase 2 of ADR-0034's hub-as-substrate rollout: the aether-hub
library crate is retired and its source moves into
aether-substrate-hub, which is now the sole home for the hub TCP
listener, engine registry, spawn supervisor, session tracking, and
MCP tool surface. Phase 1 already carried the Chassis trait impl;
this commit makes it the crate that owns the code.

The old aether-hub binary crate is gone from the workspace. Cargo
deps from aether-hub are unioned into aether-substrate-hub; no new
deps added. Internal `use aether_hub::...` references inside the
chassis wrapper (formerly lib.rs of the wrapper) become `use
crate::...`; test `use aether_hub::` become `use
aether_substrate_hub::`. eprintln log prefixes updated to
`aether-substrate-hub:` for consistency with the binary name; the
MCP-server advertised name stays `aether-hub` to keep .mcp.json
tool namespace stable.

No behaviour change: same listeners, same ports, same tests (all
pass).
