--- crates/aether-hub/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "aether-hub"
-version.workspace = true
-edition.workspace = true
-
-[dependencies]
-aether-hub-protocol = { path = "../aether-hub-protocol" }
-axum = "0.8"
-base64 = "0.22"
-postcard = { version = "1.1", default-features = false, features = ["alloc"] }
-rmcp = { version = "0.8", features = [
-    "server",
-    "macros",
-    "transport-streamable-http-server",
-] }
-schemars = "1"
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-tokio = { version = "1", features = ["rt-multi-thread", "macros", "net", "io-util", "sync", "time", "signal", "process", "fs"] }
-tokio-util = "0.7"
-uuid = { version = "1", features = ["v4", "serde"] }
-
-[target.'cfg(unix)'.dependencies]
-libc = "0.2"
-
-[dev-dependencies]
-bytemuck = { version = "1.19", features = ["derive"] }
-tokio = { version = "1", features = ["rt-multi-thread", "macros", "net", "io-util", "sync", "time"] }
