+++ crates/aether-substrate-hub/Cargo.toml
@@ -4,7 +4,27 @@ version.workspace = true
 edition.workspace = true
 
 [dependencies]
-aether-hub = { path = "../aether-hub" }
+aether-hub-protocol = { path = "../aether-hub-protocol" }
 aether-substrate-core = { path = "../aether-substrate-core" }
-tokio = { version = "1", features = ["rt-multi-thread", "macros", "signal"] }
+axum = "0.8"
+base64 = "0.22"
+postcard = { version = "1.1", default-features = false, features = ["alloc"] }
+rmcp = { version = "0.8", features = [
+    "server",
+    "macros",
+    "transport-streamable-http-server",
+] }
+schemars = "1"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+tokio = { version = "1", features = ["rt-multi-thread", "macros", "net", "io-util", "sync", "time", "signal", "process", "fs"] }
+tokio-util = "0.7"
+uuid = { version = "1", features = ["v4", "serde"] }
 wasmtime = "30"
+
+[target.'cfg(unix)'.dependencies]
+libc = "0.2"
+
+[dev-dependencies]
+bytemuck = { version = "1.19", features = ["derive"] }
+tokio = { version = "1", features = ["rt-multi-thread", "macros", "net", "io-util", "sync", "time"] }
