+++ crates/aether-substrate-hub/src/mcp/tools.rs
