+++ crates/aether-substrate-hub/tests/engine_handshake.rs
@@ -4,8 +4,10 @@
 use std::net::{Ipv4Addr, SocketAddr};
 use std::time::Duration;
 
-use aether_hub::{EngineRegistry, LogStore, PendingSpawns, SessionRegistry, run_engine_listener};
 use aether_hub_protocol::{EngineToHub, Goodbye, Hello, HubToEngine, encode_frame};
+use aether_substrate_hub::{
+    EngineRegistry, LogStore, PendingSpawns, SessionRegistry, run_engine_listener,
+};
 use tokio::io::{AsyncReadExt, AsyncWriteExt};
 use tokio::net::TcpStream;
