+++ crates/aether-substrate-hub/tests/engine_logs.rs
@@ -9,9 +9,11 @@
 use std::net::{Ipv4Addr, SocketAddr};
 use std::time::Duration;
 
-use aether_hub::{EngineRegistry, LogStore, PendingSpawns, SessionRegistry, run_engine_listener};
 use aether_hub_protocol::{
     EngineId, EngineToHub, Hello, HubToEngine, LogEntry, LogLevel, encode_frame,
+};
+use aether_substrate_hub::{
+    EngineRegistry, LogStore, PendingSpawns, SessionRegistry, run_engine_listener,
 };
 use tokio::io::{AsyncReadExt, AsyncWriteExt};
 use tokio::net::TcpStream;
